Improve tests by splitting responsibilities, using mock classes and unit tests
